DOC: Suppress warnings in visualization.rst
